pyproject.toml
[project]
name = "awesome-fastapi-ai"
version = "0.1.0"
description = "A curated collection of tools, patterns, and extensions to build FastAPI backends for AI/ML projects. Whether you're serving deep learning models or building real-time inference APIs, this repo is for you."
authors = [
    {name = "Zakariae Mrabet",email = "zakariaemrabet1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.13"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "ultralytics (>=8.3.102,<9.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "sentence-transformers (>=4.0.2,<5.0.0)",
    "langchain (>=0.3.23,<0.4.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
